feature: improve the error handling
Right now the gateway simply returns 500 errors when the incoming
request's bundle, application and event type combinations' validation
fails in the notifications backend.

The goal of these changes is to return a proper 400 response to the
client, so that they can figure out what is wrong with their payload.

Since we are at it, we also handle and log the "unreachable" errors when
the notifications backend is unreachable.

RHCLOUD-29229